pyproject.toml
[project]
name = "lookout"
version = "0.1.0"
description = "Grounded verifier for browser-use agents."
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "browser-use>=0.1.40",
    "anthropic>=0.40.0",
    "pydantic>=2.7.0",
    "jinja2>=3.1.0",
    "pillow>=10.0.0",
    "playwright>=1.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lookout"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
